requirements.txt
streamlit
python-dotenv
supabase
google-genai
langchain
langchain-google-genai
python-docx
pdfplumber
langchain-text-splitters
